+++ inst/tarpuy/setup.r
@@ -4,7 +4,7 @@
 #> open https://flavjack.github.io/inti/
 #> open https://flavjack.shinyapps.io/tarpuy/
 #> author .: Flavio Lozano-Isla (lozanoisla.com)
-#> date .: 2020-11-13
+#> date .: 2020-11-18
 # -------------------------------------------------------------------------
 
 # -------------------------------------------------------------------------
@@ -20,7 +20,6 @@ cran <- c(
   , "googleAuthR"
   , "shinydashboard"
   , "stringi"
-  , "BiocManager"
   )
 
 git <- c(